+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+# 0.7.4
+- fixed issue #112 - added keyboard shortcut to delete the current row (`ctrl+shift+alt+minus`)
+	- behaves the same was as context menu action (actually uses it)
+	- also works on mac (because normally `alt+-` [and or shift] will insert a dash)
+- added option `csv-edit.showDeleteColumnHeaderButton` true: shows a delete column button in the column header (on hover), false: not
+- added option `csv-edit.showDeleteRowHeaderButton` true: shows a delete row button in the row header (on hover), false: not
+
+
 # 0.7.3
 
 - added option `csv-edit.autoColumnWidthsIgnoreComments` to ignore comment cells for auto column sizing